requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.3
setuptools==76.1.0
